requirements.txt
pyTelegramBotAPI
mysql-connector-python
python-dotenv
tzlocal
prettytable